backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]
python-multipart
pypdf
pdfminer.six
pdfminer
chroma
langchain
langchain-community
langchainhub
bs4
ollama